requirements.txt
mock==1.0.1
nose==1.2.1
flake8==1.7.0
dnspython==1.15.0
